+++ NEWS.md
@@ -26,6 +26,8 @@
     * `rowsOrder` -> `rowOrder`
     * `UrlParameters` -> `urlParameters`
 * All the docs are now sorted properly.
+* Added `showTimerOnlyWhenLimit` to automatically hide the timer on pages that do not have a time limit.
+* Added `timeMinimum` for pages to allow for minimum time the user has to spend.
 
 ## 0.4.1